# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "description": "User Registration Endpoint",
                "operationId": "registerUser",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Registered Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Registered Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login a user",
                "description": "User Login Endpoint",
                "operationId": "loginUser",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/schools": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "summary": "Mostrar colegios",
                "operationId": "56c6b82218d9c5327f39721824360783",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los colegios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/schools/create": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "summary": "Vista Crear colegios",
                "operationId": "df857f6026e8a83961865fb02f91cb1a",
                "responses": {
                    "200": {
                        "description": "Crear todos los colegios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/schools/": {
            "post": {
                "tags": [
                    "Schools"
                ],
                "summary": "Agregar colegio",
                "operationId": "fac13e029fac3f3d526dbaf4388bf72a",
                "responses": {
                    "200": {
                        "description": "Guarda colegio."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/schools/{id}": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "summary": "Muestra un colegio por ID",
                "operationId": "1364cc874d03e93eaa23ea60e53da0e3",
                "responses": {
                    "200": {
                        "description": "Muestra colegio."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Schools"
                ],
                "summary": "Eliminar colegio",
                "operationId": "3158375b52f62cfe213b318466d571f2",
                "responses": {
                    "200": {
                        "description": "Elimina un colegio por ID."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/schools/edit": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "summary": "Vista editar colegios",
                "operationId": "54d7885ff2d2d5ee0ddc72aad41e5c98",
                "responses": {
                    "200": {
                        "description": "Editar colegio por ID."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Schools"
                ],
                "summary": "actualiza un colegio por ID",
                "operationId": "58d3d9f72dd845923888dd6f882159c9",
                "responses": {
                    "200": {
                        "description": "Actualizar colegio por ID."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Mostrar estuudiantes",
                "operationId": "7c586e93df821205d2c5af51e9dfd6d2",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los estuudiantes."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/students/create": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Vista Crear estudiantes",
                "operationId": "e918dbcaa43fb8fd3ed3b1bf8a3d8954",
                "responses": {
                    "200": {
                        "description": "Crear todos los estudiantes."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/students/": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Agregar estudiante",
                "operationId": "1f66056a9fc648f3f38e66a1948c13ca",
                "responses": {
                    "200": {
                        "description": "Guarda estudiante."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Muestra un estudiante por ID",
                "operationId": "8b4ba20f528dd1c1b78cbbb5950dfc15",
                "responses": {
                    "200": {
                        "description": "Muestra estudiante."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Eliminar estudiante",
                "operationId": "8b7dedcaa57ee3874f0220652271d1a7",
                "responses": {
                    "200": {
                        "description": "Elimina un estudiante por ID."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/student/edit": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Vista editar estudiantes",
                "operationId": "3e46734fc9fb04b6984827dc2ca6bd6e",
                "responses": {
                    "200": {
                        "description": "Editar estudiante por ID."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/students/edit": {
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "actualiza un estudiante por ID",
                "operationId": "972b68ca80ffbf43109d359eaf61983e",
                "responses": {
                    "200": {
                        "description": "Actualizar estudiante por ID."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Register",
            "description": "Register"
        },
        {
            "name": "Login",
            "description": "Login"
        },
        {
            "name": "Schools",
            "description": "Schools"
        },
        {
            "name": "Students",
            "description": "Students"
        }
    ],
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://sena.test//oauth/authorize",
                        "tokenUrl": "http://sena.test//oauth/token",
                        "refreshUrl": "http://sena.test//token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}